Some refactoring + more API implementation
