+++ core/src/main/java/org/openstack4j/model/artifact/Artifact.java
@@ -1,5 +1,9 @@
 package org.openstack4j.model.artifact;
 
+import org.openstack4j.common.Buildable;
+import org.openstack4j.model.artifact.builder.ArtifactBuilder;
+import org.openstack4j.model.common.BasicResource;
+
 import java.util.List;
 
 /**
@@ -7,7 +11,7 @@ import java.util.List;
  *
  * @author Pavan Vadavi
  */
-public interface Artifact {
+public interface Artifact extends BasicResource {
 
     String getDescription();
