+++ core/src/main/java/org/openstack4j/openstack/artifact/internal/ToscaTemplatesArtifactServiceImpl.java
@@ -2,6 +2,7 @@ package org.openstack4j.openstack.artifact.internal;
 
 import com.google.common.base.Preconditions;
 import org.openstack4j.api.artifact.ToscaTemplatesArtifactService;
+import org.openstack4j.core.transport.HttpResponse;
 import org.openstack4j.model.artifact.ArtifactType;
 import org.openstack4j.model.artifact.ArtifactUpdate;
 import org.openstack4j.model.artifact.ToscaTemplatesArtifact;
@@ -16,15 +17,19 @@ import org.openstack4j.openstack.artifact.domain.ToscaTemplatesList;
 import org.openstack4j.openstack.common.ListEntity;
 
 import java.io.File;
+import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
+
+import static org.openstack4j.core.transport.ClientConstants.CONTENT_TYPE_OCTECT_STREAM;
+import static org.openstack4j.core.transport.ClientConstants.HEADER_ACCEPT;
 
 /**
  * Created by vadavi on 19-01-2017.
  */
 public class ToscaTemplatesArtifactServiceImpl extends BaseArtifactServiceImpl implements ToscaTemplatesArtifactService {
 
-    ToscaTemplatesArtifactServiceImpl() {
+    public ToscaTemplatesArtifactServiceImpl() {
         super(ArtifactType.TOSCA_TEMPLATES);
     }
 
@@ -33,91 +38,67 @@ public class ToscaTemplatesArtifactServiceImpl extends BaseArtifactServiceImpl i
 
         return super.list(ToscaTemplatesList.class);
 
-
-        //return get(ToscaTemplatesList.class,uri("/artifacts/"+ ArtifactType.TOSCA_TEMPLATES.value())).execute();
-
     }
 
     @Override
     public ToscaTemplatesArtifact get(String artifactId) {
         Preconditions.checkNotNull(artifactId);
-        return get(ToscaTemplates.class,uri("/artifacts/"+ ArtifactType.TOSCA_TEMPLATES.value()+"/%s",artifactId)).execute();
+        return super.get(artifactId, ToscaTemplates.class);
     }
 
     @Override
     public ToscaTemplatesArtifact create(ToscaTemplatesArtifact toscaTemplatesArtifact) {
         Preconditions.checkNotNull(toscaTemplatesArtifact);
-        return post(ToscaTemplates.class,uri("/artifacts/"+ ArtifactType.TOSCA_TEMPLATES.value())).entity(toscaTemplatesArtifact).execute();
+        return super.create(toscaTemplatesArtifact, ToscaTemplates.class);
     }
 
     @Override
     public ToscaTemplatesArtifact upload(String artifactId, File file) {
         Preconditions.checkNotNull(artifactId);
-
-        Payload<?> payload = new FilePayload(file);
-
-        Invocation<ToscaTemplates> invocation = put(ToscaTemplates.class,uri("/artifacts/"+ ArtifactType.TOSCA_TEMPLATES.value()+"/%s/template",artifactId));
-        invocation.entity(payload);
-        return invocation.execute();
+        return super.upload(artifactId, file, ToscaTemplates.class, "template");
     }
 
     @Override
-    public ToscaTemplatesArtifact download(String artifactId, File file) {
-        return null;
+    public InputStream download(String artifactId) {
+        Preconditions.checkNotNull(artifactId);
+        return super.download(artifactId, "template");
     }
 
     @Override
     public ActionResponse delete(String artifactId) {
         Preconditions.checkNotNull(artifactId);
-        return deleteWithResponse(uri("/artifacts/"+ ArtifactType.TOSCA_TEMPLATES.value()+"/%s",artifactId)).execute();
+        return super.delete(artifactId);
     }
 
     @Override
     public ToscaTemplatesArtifact update(String artifactId, List<ArtifactUpdate> artifactUpdates) {
         Preconditions.checkNotNull(artifactId);
         Preconditions.checkNotNull(artifactUpdates);
-
-        Invocation<ToscaTemplates> invocation = patch(ToscaTemplates.class,uri("/artifacts/"+ ArtifactType.TOSCA_TEMPLATES.value()+"/%s",artifactId));
-        invocation.entity(new ListEntity<ArtifactUpdate>(artifactUpdates));
-        invocation.contentType("application/json-patch+json");
-        return invocation.execute();
+        return super.update(artifactId, artifactUpdates, ToscaTemplates.class);
     }
 
     @Override
     public ToscaTemplatesArtifact activate(String artifactId) {
-        return update(artifactId, "replace", "/status", "active");
+        Preconditions.checkNotNull(artifactId);
+        return update(artifactId, "replace", "/status", "active", ToscaTemplates.class);
     }
 
     @Override
     public ToscaTemplatesArtifact deactivate(String artifactId) {
-        return update(artifactId, "replace", "/status", "deactivated");
+        Preconditions.checkNotNull(artifactId);
+        return update(artifactId, "replace", "/status", "deactivated", ToscaTemplates.class);
     }
 
     @Override
     public ToscaTemplatesArtifact reactivate(String artifactId) {
-        return update(artifactId, "replace", "/status", "active");
+        Preconditions.checkNotNull(artifactId);
+        return update(artifactId, "replace", "/status", "active", ToscaTemplates.class);
     }
 
     @Override
     public ToscaTemplatesArtifact publish(String artifactId) {
-        return update(artifactId, "replace", "/visibility", "public");
-    }
-
-    private ToscaTemplatesArtifact update(String artifactId, String op, String path, String value)
-    {
         Preconditions.checkNotNull(artifactId);
-
-        ArtifactUpdateBuilder updateBuilder = ArtifactUpdateModel.builder();
-        updateBuilder.op(op);
-        updateBuilder.path(path);
-        updateBuilder.value(value);
-
-        List<ArtifactUpdate> artifactUpdates = new ArrayList<>();
-        artifactUpdates.add(updateBuilder.build());
-
-        Invocation<ToscaTemplates> invocation = patch(ToscaTemplates.class,uri("/artifacts/"+ ArtifactType.TOSCA_TEMPLATES.value()+"/%s",artifactId));
-        invocation.entity(new ListEntity<ArtifactUpdate>(artifactUpdates));
-        invocation.contentType("application/json-patch+json");
-        return invocation.execute();
+        return update(artifactId, "replace", "/visibility", "public", ToscaTemplates.class);
     }
+
 }
